+++ src/parser/parser.ts
@@ -52,6 +52,7 @@ export function parse_line(line: string, line_number: number): [boolean, string]
     if (line.length <= 0)
         return;
 
+    line = line.trimStart(); // remove initial tab
     const c = line.indexOf(" ");
     const instruction = line.substring(0, c);
     if (ISA_INSTRUCTIONS.find(x => x == instruction) == undefined || !fn_table.has(instruction)) {
@@ -60,9 +61,6 @@ export function parse_line(line: string, line_number: number): [boolean, string]
         else
             return [false, `Instruction "${instruction}" does not exist.`];
     }
-
-    //if (is_label(line))
-    //    return [true, ``];
 
     let operands_end = line.length;
     const comment = line.indexOf(";"); // ignore rest of line if comment